Open URLs with firefox by default
On Linux, the 'open' command doesn't open URLs the way it does on Mac,
so just call firefox instead. This should probably be more sophisticated
to accomodate multiple browser options, but I don't currently plan on
ever switching away from Firefox, so I'll cross that bridge if/when I
ever come to it.